Fix fitness function
For this case specifically, the only value composing a individual is the
value of x. y is obtained during the calculation of fitness function
according to the expression y = 2 - x.

This limits values obtained to be only in the domain where x + y = 2.
The fitness fucntion is calculated for x and y (obtained) and returned.